refactor(auth-service): rebranding to dialosoft
