--- src/auth-service/src/main/java/com/biznetbb/auth/persistence/entity/RefreshToken.java
@@ -1,28 +1,0 @@
-package com.biznetbb.auth.persistence.entity;
-
-import jakarta.persistence.*;
-import lombok.*;
-import org.hibernate.proxy.HibernateProxy;
-
-import java.util.Objects;
-import java.util.UUID;
-
-@Entity
-@Getter
-@Setter
-@RequiredArgsConstructor
-@AllArgsConstructor
-@Builder
-public class RefreshToken {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.UUID)
-    @Column(nullable = false, updatable = false)
-    private UUID id;
-
-    @Column(nullable = false, length = 1000)
-    private String refreshToken;
-    @OneToOne
-    @JoinColumn(name = "user_id", nullable = false)
-    private UserEntity user;
-}
