--- src/auth-service/src/main/java/com/biznetbb/auth/persistence/entity/RoleEntity.java
@@ -1,32 +1,0 @@
-package com.biznetbb.auth.persistence.entity;
-
-import com.biznetbb.auth.service.utils.RoleType;
-import jakarta.persistence.*;
-import lombok.*;
-
-import java.util.UUID;
-
-@Entity
-@Table(name = "roles")
-@Getter
-@Setter
-@RequiredArgsConstructor
-@AllArgsConstructor
-@Builder
-public class RoleEntity {
-    
-    @Id
-    @GeneratedValue(strategy = GenerationType.UUID)
-    @Column(nullable = false, updatable = false)
-    private UUID id;
-
-    @Enumerated(EnumType.STRING)
-    @Column(nullable = false, length = 50, unique = true)
-    private RoleType roleType;
-
-    @Column(nullable = false)
-    private Boolean adminRole;
-
-    @Column(nullable = false)
-    private Boolean modRole;
-}
