--- src/auth-service/src/main/java/com/biznetbb/auth/persistence/entity/UserEntity.java
@@ -1,67 +1,0 @@
-package com.biznetbb.auth.persistence.entity;
-
-import jakarta.persistence.*;
-import lombok.*;
-
-import java.time.LocalDateTime;
-import java.util.Set;
-import java.util.UUID;
-
-@Entity
-@Table(name = "users")
-@Getter
-@Setter
-@RequiredArgsConstructor
-@AllArgsConstructor
-@Builder
-public class UserEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.UUID)
-    @Column(nullable = false, updatable = false)
-    private UUID id;
-
-    @Column(nullable = false, length = 30, unique = true)
-    private String username;
-
-    @Column(nullable = false, length = 100, unique = true)
-    private String email;
-
-    @Column(nullable = false, length = 100)
-    private String password;
-
-    @Column(nullable = false)
-    @Builder.Default
-    private Boolean locked = Boolean.FALSE;
-
-    @Column(nullable = false)
-    @Builder.Default
-    private Boolean disable = Boolean.FALSE;
-
-    @ManyToMany(fetch = FetchType.EAGER)
-    @JoinTable(name = "user_roles",
-            joinColumns = @JoinColumn(name = "user_id"),
-            inverseJoinColumns = @JoinColumn(name = "role_id"))
-    private Set<RoleEntity> roles;
-
-    @Column(name = "created_at", nullable = false, updatable = false)
-    private LocalDateTime createdAt;
-
-    @Column(name = "updated_at", nullable = false)
-    private LocalDateTime updateAt;
-
-    @Column(name = "deleted_at")
-    private LocalDateTime deletedAt;
-
-    @PrePersist
-    protected void onCreate() {
-        this.createdAt = LocalDateTime.now();
-        this.updateAt = LocalDateTime.now();
-    }
-
-    @PreUpdate
-    protected void onUpdate() {
-        this.updateAt = LocalDateTime.now();
-    }
-
-}
