--- src/auth-service/src/main/java/com/biznetbb/auth/persistence/repository/UserRepository.java
@@ -1,14 +1,0 @@
-package com.biznetbb.auth.persistence.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.repository.CrudRepository;
-
-import com.biznetbb.auth.persistence.entity.UserEntity;
-
-import java.util.Optional;
-import java.util.UUID;
-
-public interface UserRepository extends JpaRepository<UserEntity, UUID> {
-    Optional<UserEntity> findByUsername(String username);
-    Optional<UserEntity> findByUsernameOrEmail(String username, String email);
-}
