--- src/auth-service/src/main/java/com/biznetbb/auth/persistence/response/JwtResponseDTO.java
@@ -1,14 +1,0 @@
-package com.biznetbb.auth.persistence.response;
-
-import lombok.*;
-
-@Getter
-@Setter
-@AllArgsConstructor
-@NoArgsConstructor
-@Builder
-public class JwtResponseDTO {
-
-    private String accessToken;
-    private String refreshToken;
-}
