--- src/auth-service/src/main/java/com/biznetbb/auth/persistence/response/ResponseBody.java
@@ -1,14 +1,0 @@
-package com.biznetbb.auth.persistence.response;
-
-import lombok.*;
-
-@Getter
-@Setter
-@AllArgsConstructor
-@NoArgsConstructor
-@Builder
-public class ResponseBody<T> {
-    Integer statusCode;
-    String message;
-    T metadata;
-}
