--- src/auth-service/src/main/java/com/biznetbb/auth/service/dto/LoginDto.java
@@ -1,9 +1,0 @@
-package com.biznetbb.auth.service.dto;
-
-import lombok.Data;
-
-@Data
-public class LoginDto {
-    private String username;
-    private String password;
-}
