--- src/auth-service/src/main/java/com/biznetbb/auth/service/dto/RegisterDto.java
@@ -1,10 +1,0 @@
-package com.biznetbb.auth.service.dto;
-
-import lombok.Data;
-
-@Data
-public class RegisterDto {
-    private String username;
-    private String email;
-    private String password;
-}
