--- src/auth-service/src/main/java/com/biznetbb/auth/service/utils/RoleInitializer.java
@@ -1,45 +1,0 @@
-package com.biznetbb.auth.service.utils;
-
-import com.biznetbb.auth.persistence.entity.RoleEntity;
-import com.biznetbb.auth.persistence.repository.RoleRepository;
-import jakarta.annotation.PostConstruct;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Component;
-
-@Component
-@RequiredArgsConstructor
-public class RoleInitializer {
-
-    private final RoleRepository roleRepository;
-
-    @PostConstruct
-    public void init() {
-
-        if (roleRepository.findByRoleType(RoleType.USER).isEmpty()) {
-            RoleEntity userRole = RoleEntity.builder()
-                    .roleType(RoleType.USER)
-                    .adminRole(false)
-                    .modRole(false)
-                    .build();
-            roleRepository.save(userRole);
-        }
-
-        if (roleRepository.findByRoleType(RoleType.ADMIN).isEmpty()) {
-            RoleEntity adminRole = RoleEntity.builder()
-                    .roleType(RoleType.ADMIN)
-                    .adminRole(true)
-                    .modRole(true)
-                    .build();
-            roleRepository.save(adminRole);
-        }
-
-        if (roleRepository.findByRoleType(RoleType.MOD).isEmpty()) {
-            RoleEntity modRole = RoleEntity.builder()
-                    .roleType(RoleType.MOD)
-                    .adminRole(false)
-                    .modRole(true)
-                    .build();
-            roleRepository.save(modRole);
-        }
-    }
-}
