--- src/auth-service/src/main/java/com/biznetbb/auth/web/config/CorsConfig.java
@@ -1,26 +1,0 @@
-package com.biznetbb.auth.web.config;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.cors.CorsConfiguration;
-import org.springframework.web.cors.CorsConfigurationSource;
-import org.springframework.web.cors.UrlBasedCorsConfigurationSource;
-
-import java.util.Arrays;
-
-@Configuration
-public class CorsConfig {
-
-    @Bean
-    CorsConfigurationSource corsConfigurationSource(){
-        
-        CorsConfiguration configuration = new CorsConfiguration();
-        configuration.setAllowedOrigins(Arrays.asList("http://localhost:8080", "http://localhost:8671")); // need eureka url
-        configuration.setAllowedMethods(Arrays.asList("POST", "DELETE", "PUT"));
-        configuration.setAllowedHeaders(Arrays.asList("*"));
-
-        UrlBasedCorsConfigurationSource source = new UrlBasedCorsConfigurationSource();
-        source.registerCorsConfiguration("/**", configuration);
-        return source;
-    }
-}
