+++ src/auth-service/src/main/java/com/dialosoft/auth/AuthApplication.java
@@ -1,0 +1,17 @@
+package com.dialosoft.auth;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
+import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
+
+@SpringBootApplication
+@EnableDiscoveryClient
+@EnableJpaRepositories
+public class AuthApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(AuthApplication.class, args);
+	}
+
+}
