+++ src/auth-service/src/main/java/com/dialosoft/auth/persistence/repository/RefreshTokenRepository.java
@@ -1,0 +1,14 @@
+package com.dialosoft.auth.persistence.repository;
+
+import com.dialosoft.auth.persistence.entity.RefreshToken;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.Optional;
+import java.util.UUID;
+
+@Repository
+public interface RefreshTokenRepository extends JpaRepository<RefreshToken, UUID> {
+    Optional<RefreshToken> findByRefreshToken(String token);
+    Optional<RefreshToken> findByUserId(UUID userId);
+}
