+++ src/auth-service/src/main/java/com/dialosoft/auth/persistence/repository/UserRepository.java
@@ -1,0 +1,12 @@
+package com.dialosoft.auth.persistence.repository;
+
+import com.dialosoft.auth.persistence.entity.UserEntity;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+import java.util.UUID;
+
+public interface UserRepository extends JpaRepository<UserEntity, UUID> {
+    Optional<UserEntity> findByUsername(String username);
+    Optional<UserEntity> findByUsernameOrEmail(String username, String email);
+}
