+++ src/auth-service/src/main/java/com/dialosoft/auth/persistence/response/JwtResponseDTO.java
@@ -1,0 +1,14 @@
+package com.dialosoft.auth.persistence.response;
+
+import lombok.*;
+
+@Getter
+@Setter
+@AllArgsConstructor
+@NoArgsConstructor
+@Builder
+public class JwtResponseDTO {
+
+    private String accessToken;
+    private String refreshToken;
+}
