+++ src/auth-service/src/main/java/com/dialosoft/auth/persistence/response/ResponseBody.java
@@ -1,0 +1,14 @@
+package com.dialosoft.auth.persistence.response;
+
+import lombok.*;
+
+@Getter
+@Setter
+@AllArgsConstructor
+@NoArgsConstructor
+@Builder
+public class ResponseBody<T> {
+    Integer statusCode;
+    String message;
+    T metadata;
+}
