+++ src/auth-service/src/main/java/com/dialosoft/auth/service/RefreshTokenService.java
@@ -1,0 +1,69 @@
+package com.dialosoft.auth.service;
+
+import com.dialosoft.auth.persistence.entity.RefreshToken;
+import com.dialosoft.auth.persistence.entity.UserEntity;
+import com.dialosoft.auth.persistence.repository.RefreshTokenRepository;
+import com.dialosoft.auth.web.config.jwt.JwtUtil;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.stereotype.Service;
+
+import java.util.Optional;
+import java.util.UUID;
+
+@Service
+public class RefreshTokenService {
+
+    @Autowired
+    private RefreshTokenRepository refreshTokenRepository;
+    @Autowired
+    private UserSecurityService userSecurityService;
+    @Autowired
+    private JwtUtil jwtUtil;
+
+    public RefreshToken getOrCreateRefreshTokenByUserName(String username) {
+
+        UserEntity userEntity = userSecurityService.getUserByUserName(username);
+
+        return refreshTokenRepository.findByUserId(userEntity.getId())
+                .filter(this::isRefreshTokenValid)
+                .orElseGet(() -> createRefreshToken(userEntity));
+    }
+
+    private RefreshToken createRefreshToken(UserEntity userEntity) {
+
+        UserDetails userDetails = userSecurityService.loadUserByUsername(userEntity.getUsername());
+        String stringRefreshToken = jwtUtil.generateRefreshToken(userDetails.getUsername(), userDetails.getAuthorities());
+
+        RefreshToken refreshToken = RefreshToken.builder()
+                .user(userEntity)
+                .refreshToken(stringRefreshToken)
+                .build();
+
+        return refreshTokenRepository.save(refreshToken);
+    }
+
+    public Optional<RefreshToken> findByToken(String token){
+        return refreshTokenRepository.findByRefreshToken(token);
+    }
+
+    public Optional<RefreshToken> findByUserId(UUID userId){
+        return refreshTokenRepository.findByUserId(userId);
+    }
+
+    public RefreshToken verifyRefreshTokenExpiration(RefreshToken refreshToken) {
+        if (!isRefreshTokenValid(refreshToken)) {
+            throw new RuntimeException(refreshToken.getRefreshToken() + " Refresh token is expired. Please login again.");
+        }
+        return refreshToken;
+    }
+
+    private boolean isRefreshTokenValid(RefreshToken refreshToken) {
+        if (!jwtUtil.isValid(refreshToken.getRefreshToken())) {
+            refreshTokenRepository.delete(refreshToken);
+            return false;
+        }
+        return true;
+    }
+
+}
