+++ src/auth-service/src/main/java/com/dialosoft/auth/service/dto/LoginDto.java
@@ -1,0 +1,9 @@
+package com.dialosoft.auth.service.dto;
+
+import lombok.Data;
+
+@Data
+public class LoginDto {
+    private String username;
+    private String password;
+}
