+++ src/auth-service/src/main/java/com/dialosoft/auth/service/dto/RefreshTokenDto.java
@@ -1,0 +1,14 @@
+package com.dialosoft.auth.service.dto;
+
+import jakarta.validation.constraints.NotNull;
+import lombok.*;
+
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class RefreshTokenDto {
+
+    @NotNull(message = "Refresh token is required.")
+    private String refreshToken;
+}
