+++ src/auth-service/src/main/java/com/dialosoft/auth/service/dto/RegisterDto.java
@@ -1,0 +1,10 @@
+package com.dialosoft.auth.service.dto;
+
+import lombok.Data;
+
+@Data
+public class RegisterDto {
+    private String username;
+    private String email;
+    private String password;
+}
