+++ src/auth-service/src/main/java/com/dialosoft/auth/service/utils/RoleType.java
@@ -1,0 +1,15 @@
+package com.dialosoft.auth.service.utils;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public enum RoleType {
+
+    USER("ROLE_USER"),
+    ADMIN("ROLE_ADMIN"),
+    MOD("ROLE_MOD");
+
+    private final String roleName;
+}
