+++ src/auth-service/src/main/java/com/dialosoft/auth/web/config/jwt/JwtFilter.java
@@ -1,0 +1,67 @@
+package com.dialosoft.auth.web.config.jwt;
+
+import java.io.IOException;
+
+import org.springframework.http.HttpHeaders;
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.User;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.web.authentication.WebAuthenticationDetailsSource;
+import org.springframework.stereotype.Component;
+import org.springframework.web.filter.OncePerRequestFilter;
+
+import jakarta.servlet.FilterChain;
+import jakarta.servlet.ServletException;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+
+@Component
+public class JwtFilter extends OncePerRequestFilter {
+
+    private final JwtUtil jwtUtil;
+
+    private final UserDetailsService userDetailsService;
+
+    public JwtFilter(JwtUtil jwtUtil, UserDetailsService userDetailsService){
+        this.jwtUtil = jwtUtil;
+        this.userDetailsService = userDetailsService;
+    }
+
+    @Override
+    protected void doFilterInternal(
+        HttpServletRequest request,
+        HttpServletResponse response,
+        FilterChain filterChain
+        )   throws ServletException, IOException {
+        
+        String authHeader = request.getHeader(HttpHeaders.AUTHORIZATION);
+
+        if(authHeader == null || authHeader.isEmpty() ||!authHeader.startsWith("Bearer")){
+            filterChain.doFilter(request, response);
+            return;
+        }
+
+        String accessJwtToken = authHeader.split(" ")[1].trim();
+        if(!this.jwtUtil.isValid(accessJwtToken)){
+            filterChain.doFilter(request, response);
+            return;
+        }
+        
+        String username = this.jwtUtil.getUsername(accessJwtToken);
+        User user = (User)this.userDetailsService.loadUserByUsername(username);
+
+        UsernamePasswordAuthenticationToken authenticationToken = new UsernamePasswordAuthenticationToken(
+            user.getUsername(),
+            user.getPassword(),
+            user.getAuthorities());
+        
+        authenticationToken.setDetails(
+            new WebAuthenticationDetailsSource()
+            .buildDetails(request)  
+        );
+
+        SecurityContextHolder.getContext().setAuthentication(authenticationToken);
+        filterChain.doFilter(request, response);
+    }
+}
