--- src/auth-service/src/test/java/com/biznetbb/auth/AuthApplicationTests.java
@@ -1,15 +1,0 @@
-package com.biznetbb.auth;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.ActiveProfiles;
-
-//@SpringBootTest
-//@ActiveProfiles("test")
-//class AuthApplicationTests {
-//
-//	@Test
-//	void contextLoads() {
-//	}
-//
-//}
