+++ src/auth-service/src/test/java/com/dialosoft/auth/AuthApplicationTests.java
@@ -1,0 +1,15 @@
+package com.dialosoft.auth;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.ActiveProfiles;
+
+//@SpringBootTest
+//@ActiveProfiles("test")
+//class AuthApplicationTests {
+//
+//	@Test
+//	void contextLoads() {
+//	}
+//
+//}
